do not remove binding attributes
not removing the data-bind attributes causes an error in
foreach bindings. they are visited twice, and the second time
around, htmlizer will attempt to apply the bindings against
the parent context, instead the foreached list it did in the
first time. that second pass will either cause very weird bugs
or it will not work at all.